Update patch set 3
Patch Set 3:

> https://git.eclipse.org/r/#/c/150030/ is submitted but the same
 > tests still fail reproducibly

RebaseOperationTest.testStopAndAbortOnConflict also fails for me locally. The problem is that Project-1 is not shared with EGit. So either the test setup is wrong, or we should change the ResourceRefreshHandler to not use ResourceUtil.getContainerForLocation() in lines 173 and 189 but a similar variant that would do the same, but not check whether the container is shared with git.

Patch-set: 3
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0