Don't return a path for an IStorage representing an index entry
Doing so may make some parts of Eclipse conclude the storage really
represented an on-disk file and then revert the internal file buffer
to that index state, which rolls back all open editors on that file
to the index state (and marks them dirty). That's not what we want
at all. Occurs for instance with Java files.

Adapting to IFile (from commit 0c17469b) to make Eclipse select the
right DocumentProvider is fine, but returning a path for the IStorage
is not.

Bug: 550989
Change-Id: I3f63ad9a4432063d7d1a245c05641fcf12698c2f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>