Update patch set 3
Patch Set 3: Code-Review-1

This looks better than the previous implementation in the sense that the checkbox reacts to selection handling (it has a blue border instead of a black one when hovering the checkbox) and it works well for the checkbox as such.
However, I found one quirk: When clicking the source ref cell, the full row is selected and each cell has a background color. Clicking the source ref cell of another row, this background color is removed again everywhere, but not in the area of the checkbox. I have to hover on the checkbox to make it repaint its background with the "no selection" background color.
I'll check if I find a fix for that...

Patch-set: 3
Label: Code-Review=-1