Update patch set 3
Patch Set 3:

(1 comment)

As there were so many comments, I wanted split the patch. I think that I have dealt with everything but the plugin.xml (default handler etc. which is the must difficult for me) and the image file (which I don't know how to deal with).

I don't know if it is a problem, that I rebased on master, hopefully not.

With respect to drag'n'drop. Even if there are many repositories (or the view is very small), (for me) there is always a very small area after the last element of the tree. I can drop a repository from a group there.

I am still not sure about showing a repository on drop. The current behaviour seems to be consistent with what happens if you drop a file in a folder in the package explorer - the file is moved, the folder remains closed.

Patch-set: 3