Update patch set 3
Patch Set 3:

> ...
 > repetitive, I was looking at the property testers.

Stick to the expression language in XML where it fulfills the needs, even if it is verbose. No need to duplicate core functionality in custom property testers.

There is already the RepositoriesViewPropertyTester. Perhaps add new property tests, if needed, in there.

 > probably need quite a few new properties:
 > e.g. onlyReposSelected, onlyGroupsSelected, onlyReposInGroupsSelected
 > (and maybe for later onlyVisibleGroupsSelected, onlyHidableGroupsSelected).

The first two should not be done via property testers; the Eclipse expression language is completely sufficient for those.

Patch-set: 3
Label: Verified=0