Update patch set 2
Patch Set 2:

The JavaProjectTester also runs project creation and deletion in workspace operations. Probably TestProject should also do so. I don't remember what the purpose of the TestUtil.waitForJobs() calls in JavaProjectTester was. Another difference is that the JavaProjectTester also makes sure that the project is shared with EGit.

When I wrote these Java tests in UI I hadn't realized that the core TestProject also created a Java project...

Patch-set: 2