Update patch set 5
Patch Set 5: Code-Review-2

Nope, sorry. Needs a rework; there's something I don't understand.

Up to PS 3, clicking the button would make the view automatically show the currently selected element *in the previously active view*. And in fact, one test now actually tests for this and passes because it just changes the state, but doesn't click the button anymore.

But the code cannot work. The active part when that command is fired through clicking the button is the repo view itself. And indeed when I manually click the button, the view does not follow the selection! It can't, since the active part is already the view itself. if we'd want that behavior, we'd have to track the "previously current" selection through a PartListener.

So it looks like I at least I should change the tests back to actually clicking the button. And the view will react only when the user then activates some other part.

Patch-set: 5
Label: Code-Review=-2