Update patch set 6
Patch Set 6: -Code-Review

Test changed back to actually clicking that button. (For the one test where it matters.)

The view not reacting immediately based on the selection in the previously active part is actually bug 409722. This is to be fixed in a separate commit.

Patch-set: 6
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0