[repo view] Link with selection: immediate reaction, show open editors
React immediately when linking with selection is enabled. Track
workbench parts even when linking is disabled. When linking
is enabled, immediately react on the last selection before the
repositories view became active.

When the selection inside the view changes react to a FileNode
being selected by bringing an already open editor of that file
to the front.

Activate and correct the already existing test for this.

Remove the old LinkHelper. It wasn't used anymore since commit
d6cf5241 in 2013.

Bug: 409722
Change-Id: I49ebc0b361469f3058348199ff8bd3557d949634
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>