Update patch set 8
Patch Set 8:

Thanks for the additional feedback. I rebased to the current master and dealt with all review comments for Patch Set 7. Command descriptions were added, labels adapted, etc.

I checked the callers of RepositoryTreeNode.getRepository(). In most cases, a null check was already present or not necessary as the relevant node types do not return null. In one case, I added a null check. The current state looks OK.
I was not aware that returning null here might be such a dramatic change: In several places I came across so far, there were null checks already (in particular RepositoryTreeNode#getPath right below the actual getRepository method). So, I assumed that there were existing cases where the repository is null.

Most of the repository view commands depend on the (single) repository connected to the tree node. I saw a feature request for "batch" commands (same command executed on multiple selected repositories). If there are plans to implement something like that, group nodes might always have to be handled separately (unless the API is changed from RepositoryTreeNode.getRepository to RepositoryTreeNode.getRepositories).

(Note also, that the current "batch" commands Pull, Swith To and Garbage Collect are not visible when a group is selected. But I would not like to deal with this "bug" in the current context. I'd like to get the first version up and running.)

Patch-set: 8
Label: Verified=0