Update patch set 13
Patch Set 13:

Great :-)

One thing that totally confused me first: the drop down and "Ref Filters..." menu should be disabled if we are in the "single branch" mode, otherwise user sees in the dialog configuration, changes something, but nothing happens.

Same with the icon: should be clear that we are in the single vs multiple branch modes.

The rest is genius!

Cool!
Can't wait to get this merged!

Patch-set: 13
Label: Code-Review=0