Update patch set 10
Patch Set 10:

I have addressed the comments whose fixes were obvious to me.
Summary of unclear points:
* command label
* (rename command)
* relative repo path in preferences

> 1. Do we really want to allow dragging a submodule into a repo group? 
No. I have never worked with submodules and was not aware of the previous behaviour. I added guards for not adding those repositories to a group and for not showing the groups sub menu.

> 2. All menu commands that open a dialog should have an ellipsis ... at the end of the label.
I assume, you are referring to "Create a Repository Group". This entry appears in a tree node's context menu and in the view's main menu. In the view's menu, the add, create and clone repository entries are without ellipes although a dialog is opened. So, I left it as well for consistency.
Please specify, which of the two entries (or both) should be changed.

> 3. Can we tie the "Rename Repository Group" to the standard rename command?
Not easily. If the standard rename command is used, the key binding F2 for renaming a group is shown in the context menu (which is nice), but it does not work, because Egit defines a separate keybinding for the rename branch command (org.eclipse.egit.ui.team.RenameBranch) in the view's context, overriding the default key binding. I don't know why there are two command ids for renaming a branch.

I assume "Move to <groupName>" in the group menu of a regular repository is OK. Michael Keppler asked me to shorten "Move to group..." in a previous comment. For submodules the menu is not shown anymore.

Using RepositoryUtil#relativizeToWorkspace in RepositoryGroups naively does not work. The compare/remove operations for determining whether a repository belongs to a group fail as a consequence. The current implementation implicitly uses a "normalized" file due to RepositoryUtil#getRepositories. The inverse operation for relativizeToWorkspace (RepositoryUtil line 472) or the workspacePath are currently not public but would be needed for instantiating the repository groups from the preferences.
Should a method public Path getAbsolutePath(String repositoryDir) be added to RepositoryUtil or should I rather duplicate the workspacePath code in RepositoryGroups? I ask, because I am reluctant to add API to such a central component.

Patch-set: 10
Label: Verified=0