Update patch set 1
Patch Set 1:

Shouldn't the display label strings be in the UI plugin and the label provider should decide which string to use for which item? Perhaps we have other such cases, but I find it strange that some enum in the core plugin should have to worry about human-readable labels.

Patch-set: 1