Update patch set 1
Patch Set 1:

Well, I'm not a fan of defining an enum somewhere, and then a "wrapper" for the enum somewhere else, just to add one more field to it (or even worse, a switch case over the enum and returning the label instead). If I rename "label" to "name", and create a method getLabel() in the UI plugin, which says "return action.name;" would you be more happy in that case? Cause that is basically exactly what we had before...

Patch-set: 1
Reviewer: Gerrit User 440 <440@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0