Update patch set 1
Patch Set 1:

I think the choice of name() before was wrong; we just called some method on that object that happened to return a String. Could have used toString(), too. If we had a label provider somewhere that used Repository.toString() to get a label, would you then add a getLabel() method to Repository? Or take Eclipse Resources: labels are provided in UI label providers (typically through adapters in that case), not by the Resource.

Patch-set: 1
Label: Code-Review=0