Update patch set 2
Patch Set 2:

These gitflow tests create a git repo inside the project directory. Probably that contributes to the problems with deleting the projects. It also connects the project to the repo, but doesn't disconnect the project at the end. The GitMoveDeleteHook forbids deleting folders that are still connected.

If that is the problem I don't understand why these tests can work at all. We also should get a message somewhere that there was an active git repository in a folder. But in any case those test setups look fishy.

Patch-set: 2