Update patch set 20
Patch Set 20:

> > Wasn't this limit increased to some higher value recently? Or was
 > > it for committers only?
 > 
 > Not that I know of. The Genie also commented; see second comment
 > above. For committers, there is no limit.

The limit was increased to 1000 lines a couple of years back. Above 1000 LoC still a CQ is required, see [1]

The requirement for CQs for 3rd party dependencies was relaxed with the new version of the IP policy [2] which was approved by the board of directors of the Eclipse Foundation. There will be a tool [3] to check in the database of clearly defined [4] if licenses in the version of a 3rd party component complies to the list of allowed licenses. If this check succeeds no CQ is required. 3rd party dependencies implementing security stuff like encryption still need a CQ. Though this tool isn't yet ready and the process update not yet documented hence it's currently not completely clear how to proceed. If in doubt email Wayne.

[1] https://www.eclipse.org/projects/handbook/#ip-project-code
https://www.eclipse.org/legal/EclipseLegalProcessPoster.pdf
[2] https://www.eclipse.org/org/documents/Eclipse_IP_Policy.pdf
[3] https://blogs.eclipse.org/post/wayne-beaton/revising-eclipse-ip-policy-third-party-content
[4] https://clearlydefined.io/

Patch-set: 20
Label: Verified=0