Update patch set 1
Patch Set 1:

I do see the decorator disappearing and reappearing, but I have no idea how to fix that correctly. I did a little bit of debugging (well, I inserted some sysouts).

a) There are quite a few calls to getImage when performing the refresh, one when creating the tree item and two updates to the same item afterwards. I don't know if they are all necessary.
b) More interestingly, it seems to be a race condition - at least in the particular case of the GitFlowDecorator.

What I did was insert a sysout in GitFlowLightweightDecorator#decorateIcons and one in my modified RepositoryTreeNodeLabelProvider#getImage. In the latter, I stored the image from the super call in a variable and printed its hash code to the console before returning it.

What I noticed was that without setting a breakpoint, the first image (create tree item) was obtained *before* the git flow decoration was invoked, so the hash code of the first image was different from the next two invokatons (after decoration was invoked).

When I set a breakpoint in the getImage-Method, the git flow decoration is invoked before the first image is obtained and all three images have the same hashcode. Maybe LightweightDecoration is not the right choice here!?

Patch-set: 1
Reviewer: Gerrit User 236689 <236689@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0