Update patch set 1
Patch Set 1:

Not a race condition, unfortunately. The first time, no decoration is computed, so the decorating label provider just returns the unadorned base image and queues a decoration request. This request is handled in the background vbia the lightweight decorator, and once done, the an event is fired that causes the viewer to refresh the label. It calls the decorating label provider again, which gets the image again, but this time the decorations are there, and thus it returns the decorated image.

All normal. And using a LightweightDecorator _is_ the way to go -- we really don't want to compute all this stuff on the UI thread.

I haven't figured out yet how package explorer or project explorer manage to avoid that blinking effect. Either we're doing something wrong somewhere, or it doesn't occur for them because of using the Eclipse resource tree and ResourceChangedEvents, or their decorators are just so fast one doesn't notice it. Another speciality of the RepositoriesView is that it schedules its own refresh job -- maybe the perceptible delay until the decorations are there again comes from that.

I'm working on making these decorators faster anyway; I may revisit this once I have that done. But if you find a solution before that (and that doesn't move even more into the UI thread) it'd be OK, too.

Patch-set: 1
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0