Update patch set 12
Patch Set 12:

I noticed that I did not update all the copyright headers. Also, I further added the branch name as a group decorator in case there is a unique branch. I did not change the label provider in the group delete confirm dialog as was done in change 151653.

I hope, these additional (non-)changes are OK.

Thanks for the patience and the thorough reviews!

Patch-set: 12
Label: Verified=0