Update patch set 13
Patch Set 13: Code-Review+1

(1 comment)

> further added the branch name as a group decorator in case there is
 > a unique branch.

I had that idea, too, but thought it was something for a follow-up change.
Can be improved further in a follow-up commit: on Working Sets in the package explorer, we show "[* branch]" if the working set contains projects from several repos all on the same branch. For repo groups that have more than one repo we could also show this asterisk.

 > I hope, these additional (non-)changes are OK.

Sure.

Now we'll just have to wait for the CQ.

Patch-set: 13