[repo view] Bind "Rename Repository Group..." to standard rename
Use the standard rename for this command handler, and also for the
"Rename Branch..." when a Ref node is selected. These show up in
Quick Access as "Rename the selected item...", which makes sense,
and are bound to F2.

Use the special RenameBranch command only when a repository node
is selected. Add a key binding for F2 in a special context, and
activate it only when a single repository is selected to avoid
conflicts with the standard binding. This command shows up in
Quick Access as "Rename Branch...".

Change-Id: If95b79e83436afaa956593f0c971552eeddc5441
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>