Comparisons of submodules: fix and make non-editable
Invoking "Compare with index" in the staging view on a submodule
didn't work. It showed an empty editable index part, and when edited
and saved would stage a file.

Show the correct content: the object ID from the index is the HEAD
of the submodule and is also the content. There is no blob with this
ID.

Forbid editing such index entries as it's much too fragile and can
lead to corrupt data far too easily.

In LocalNonWorkspaceTypedElement also fix finding the submodule
repository; in some cases the former code could not find it and
then tried to open an InputStream on the directory.

Bug: 498547
Change-Id: I1c7a18dba72d69068620bf8ff4c95011006c0df7
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>