[repo groups] inline renaming of repository groups
Set up editing support in the repositories view, triggered
programmatically via the RenameRepositoryGroup handler. Fix
the RepositoriesViewSorter to actually consider the categories,
and ensure repository groups are at the bottom.

Fix equals(), hashCode(), and compareTo() for RepositoryGroupNodes:
these should store the RepositoryGroup, not the label as their object.
Remove the then superfluous getGoup() method; use getObject() instead.
equals() must compare the repo groups. compareTo() is used only for
sorting, so it may compare by group name.

Fix the label provider: the workbench adapter must return a label
text, which is the group's name.

Also eliminate the default handler for PasteCommand by tying the
handler to the standard org.eclipse.ui.edit.paste command.

Pasting proved to be a problem since the view registered a default
handler for the paste command, which "stole" the paste and always
handled it. Thus pasting into an open cell editor was not possible.

JFace does have some support for overriding edit _actions_ when a
cell editor is active, but I haven't found anything similar for
commands. In the end I settled for making the PasteCommand simply
call the view first to give it a chance to forward the paste to the
cell editor.

Change-Id: If1cda3768bb284d1a6801ec63dfaa80cb684e940
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>