[repo groups] move to own package
Repository groups are in fact a general concept, not tied to the
repositories view. (Even though currently we use them only there,
we will discover that in order to support bulk operations we'll
reference them from other general code, too.)

Change UIText keys accordingly, and drop the "GitRepositoriesView"
prefix in the preference keys.

Change-Id: I9df4873581d7a6e15ca2200caafee2b1cec6bb49
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>