[repo groups] "Don't show again" for deletion confirmation
Introduce a preference for showing this confirmation dialog (by
default true). Add a checkbox to the dialog to toggle the setting.
Add an item on the "Confirmations & Warnings" preference page to
set and re-set the setting.

Show the dialog only if the setting is true, and there are any
repositories in the to-be-deleted repository groups.

Change-Id: Ia4b292af2fd84844a3f4c64f7a3d97db9c72e6ff
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>