Update patch set 2
Patch Set 2:

I'd like to comment on this change in general. My feeling is that rather than introducing a special context for renaming groups, there should be a special context for renaming the branch when the repository is selected.

Renaming a group is a standard use case (rename the element that is actually selected) and could actually have the standard UI's rename command ID (including default key binding). Think of a potential new feature of renaming a remote. You would have to make changes for every such "rename acutal resource" command.

Patch-set: 2
Label: Code-Review=0