Update patch set 4
Patch Set 4:

(6 comments)

Changing RepositoryGroupNode should definitely be done! Note that the javadoc for RepositoryTreeNode#getObject has to be adapted as well.

Sorry, that I am not immediately ecstatic about the inline renaming. The reason is not that the feature itself is not nice. It is just that a) currently it is not consistent with the other renaming actions in the view - they all open a dialog, b) the validation is implemented in two places - and you always have to think of both of them in case of changes.

Is there a plan for changing renaming a branch inline as well? The different modes may not be intuitive to the user.
If the inline renaming should win, how about creating a group with a default name and immediately invoke the rename command (or with an empty name and deleting it if renaming is aborted). This way, there would not be two different versions for defining a group name.

Patch-set: 4
Label: Verified=0