Update patch set 4
Patch Set 4:

(2 comments)

> (6 comments)
 > 
 > Changing RepositoryGroupNode should definitely be done! Note that
 > the javadoc for RepositoryTreeNode#getObject has to be adapted as
 > well.
 > 
 > Sorry, that I am not immediately ecstatic about the inline
 > renaming. The reason is not that the feature itself is not nice. It
 > is just that a) currently it is not consistent with the other
 > renaming actions in the view - they all open a dialog, b) the
 > validation is implemented in two places - and you always have to
 > think of both of them in case of changes.

I know about (a) and (b). But both would need to be done in follow-up changes.

 > Is there a plan for changing renaming a branch inline as well? The
 > different modes may not be intuitive to the user.

I thought about it, and at least for renaming on a Ref node it could be easily done -- except that the current dialog also offers content assist (when the user enters invalid characters. So we'd need to have a way to display an error message and the content assist without having them overlap...

For renaming the currently checked out branch when invoked on the repo node itself we still would need the dialog. But I think a user would understand that, since in that case it's _not_ the selected object that is to be renamed.

 > If the inline renaming should win, how about creating a group with
 > a default name and immediately invoke the rename command (or with
 > an empty name and deleting it if renaming is aborted). This way,
 > there would not be two different versions for defining a group
 > name.

Yes, for creation I had that idea, too. But that would be a follow-up change.

Patch-set: 4