Update patch set 7
Patch Set 7: Code-Review-2

Besides my forgetting to update the test I broke something else. The rename per se works, but there's something mighty strange going on between the staging view and the repo view. The staging view creates a RepositoryNode for its title, and nulls that node out via clear() when it no longer needs it. But somehow that very node instance ends up replacing the different instance in the repo view tree?!

So as long as one only selects repositories in the repo view (with staging set to follow the selection) everything's fine. But as soon as one selects a repo group, the staging view gets a null repo, discards and clears its title node, and since that same instance somehow is also referenced in the repo view's tree, we now have a repo node with a null repo in the repo view, wich leads to numerous problems.

Investigating how that staging view RepositoryNode ends up being in the repo view tree, but that may take some time.

Patch-set: 7
Label: Code-Review=-2