Update patch set 13
Patch Set 13: Code-Review-1

There's something mighty strange going on between the staging view and the repo view. The staging view creates a RepositoryNode for its title, and nulls that node out via clear() when it no longer needs it. But somehow that very node instance ends up in the repo view tree?!

So as long as one only selects repositories in the repo view (with staging set to follow the selection) everything's fine. But as soon as one selects a repo group, the staging view gets a null repo, discards and clears its title node, and since that same instance somehow is also referenced in the repo view's tree, we now have a repo node with a null repo in the repo view, which leads to numerous problems.

Something's really fishy.

Steps to reproduce:

1. Start Eclipse. Mine has both repo view and staging view visible, staging view set to follow selection, repo view not.
2. I have multiple repos in the repo view, and some groups.
3. Click on a repo. Staging view updates.
4. Click on a repo group. Staging view updates, clears and says "No repository".
5. Click on the same repo as in step 3.

Expected: staging view updates and shows the repo name in its title.

Actual: since the node clicked in step 5 has myRepository == null (cleared by the staging view in step 4 since somehow it's the same instance), nothing happens. Git operations on that node are broken and cause exceptions.

How the heck does the staging view's RepositoryNode end up in the tree in the repo view?

@Alexander: can you reproduce?

(I discovered that in https://git.eclipse.org/r/#/c/151813/ and thought I had broken it somehow, but it already occurs in this version.)

Patch-set: 13