Update patch set 13
Patch Set 13:

Occurs on two different computers for me (both with a child Eclipse using the Neon target). So it's not a corrupted workspace.

To be clear: I don't think this instance sharing is introduced by this change. It probably happened before, but we never had a RepositoryTreeNode with a null repository until this change, and so this never caused problems. Of course it could be "fixed" in staging view in several ways (not reacting on the selection if repo == null, or not clearing its titleNode), but that doesn't solve the fundamental problem: why does that titleNode end up being referenced from the repo view's tree?? That should just not happen!

Patch-set: 13