Update patch set 13
Patch Set 13: -Code-Review

OK, found it: bug 552622. Nasty. Probably undetected until now because it's actually difficult to give the staging view a null repository so that it would clear out its titleNode. Apparently nobody ever closes the staging view :-), because then it also clears that node and one has the same problems. Even without this change.

With this change it's become super-easy to give the staging view a null repo and make it clear its titleNode: just select a repo group.

I'm still fine with this change. I'll push a fix for bug 552622, and then I'll rebase this whole sequence to pick up this fairly crucial fix.

Patch-set: 13
Label: Code-Review=0