Update patch set 14
Patch Set 14:

Rebased so that we get bug 552622 out of the way.

Also fixed copyright headers. We use "Copyright (C) FROM, TO Author", where FROM, TO are two years and give a range. (Yes, personally I would have written a range with a dash, but that's just the format EGit uses.) TO can be omitted for new files, of course. And it's not 100% consistent in the EGit sources. Generally I correct ones I see with more than two years.

Patch-set: 14
Label: Verified=0