[history] show new HEAD or FETCH_HEAD in commit table
Commit 9a0e6920 avoided jumping around in the git history when the
branches to show changed. This was done by trying to keep the top
commit in the commit graph table the same.

This is nice when only the display settings change, but it also
meant that the table would not scroll to show newly fetched commits,
or when a branch switch occurred. It's a bit disconcerting to fetch
some changes from Gerrit or switch branches and then not see in the
history view what was fetched or where one switched to.

Force the commit table to show the new HEAD or FETCH_HEAD if either
changed. (If only FETCH_HEAD changes this currently works only on
"Fetch from Gerrit" since JGit doesn't notify when FETCH_HEAD changes.
See bug 551243.)

Change-Id: Ib65687411dad419f4406f9053b3d4812a6dde956
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>