Cache the repository config
Many JGit operations on repositories need to get values from the
repository config. Each time, JGit check whether the config or
any of its parent config have changed and reloads them, if needed.

Even if no reload is required the mere checking whether the file
was modified can be an expensive operation, especially since it
also checks the user and system config. At least in corporate
Windows setups it's not uncommon to have the home directory actually
on a share, i.e. a network disk. In that case EGit performance
plummets; many operations and especially UI updates can become
extremely sluggish. EGit makes thousands of calls that access the
repo config, and slow access to the filesystem the user config makes
the UI barely useable.

Introduce the concept of a "unit of work". Within a unit of work the
git configs for a repository are checked and possibly loaded only
exactly once, and subsequent operations work with these cached values
until the end of the unit. Caching of repository configs is per thread.

Add a CachingRepository subclass of FileRepository that supports this
via ThreadLocals. Since the user's only way to influence this caching
is via UnitOfWork, it is assured that nothing is leaked through these
ThreadLocals.

Use the new UnitOfWork in the repository state caches, when decorating,
and when computing IndexDiffs.

Bug: 456996
Change-Id: I751b7c6c5ff02d387975beb6a5076f200ff19760
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>