[repo groups] inline renaming of repository groups
Set up editing support in the repositories view, triggered
programmatically via the RenameRepositoryGroup handler.

Also use inline renaming for creating groups: create the group with
a unique name, then immediately start the rename process. If the
user then cancels the name edit, there will be a new group with
the unique default name. This behavior is consistent with the way
folders are created in many graphical UIs, for instance on Windows
or also on Mac.

Remove the ellipsis from the menu items since the commands now don't
open dialogs anymore.

Pasting proved to be a problem since the view already has a paste
command, which "stole" the paste and always handled it. Thus pasting
into an open cell editor was not possible.

JFace does have some support for overriding edit _actions_ when a
cell editor is active, but I haven't found anything similar for
commands. In the end I settled for making the PasteCommand simply
call the view first to give it a chance to forward the paste to the
cell editor.

Change-Id: If1cda3768bb284d1a6801ec63dfaa80cb684e940
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>