[repo groups] Inline renaming of repository groups
Set up editing support in the repositories view, triggered
programmatically via the RenameRepositoryGroup handler.

Pasting proved to be a problem since the view already has a paste
command, which "stole" the paste and always handled it. Thus pasting
into an open cell editor was not possible.

JFace does have some support for overriding edit _actions_ when a
cell editor is active, but I haven't found anything similar for
commands. In the end I settled for making the PasteCommand simply
call the view first to give it a chance to forward the paste to the
cell editor.

Change-Id: If1cda3768bb284d1a6801ec63dfaa80cb684e940
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>