Update patch set 25
Patch Set 25:

> It is because it'll just stop after 5 seconds. May take longer on
 > Jenkins.

bot.waitUntil will also timeout at some point (default is also 5 seconds, but I don't know if the default is used here).

Andreys suggestion was to increase the timeout for waitForJobs to something like 30 seconds as well as adding a log message to waitForJobs to log when the timeout is hit.

One more advantage of your version I just noticed is that there is no need to wait an arbitrary time (50ms) for the job to enter the queue.

I'm not sure which version is really the better solution.

Patch-set: 25
Label: Verified=0