Defer subscriber initialization to background job
The initialization of the subscriber is a potentially long running task.
The GitRepositoryProviderType is instantiated on the main thread, thus
the UI is blocked on startup until the subscriber was initialized.

This change encapsulates the subscriber initialization into a background
system job, so that the main thread is no longer blocked. The
getSubscriber() method synchronizes on the job if it was not already
finished.

Bug: 552930
Change-Id: Id26f79ba7c2aa26be39e1a4f44ce7f8e7a6188a9
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>