[history] Alignments in the ref filter dialog
For the left alignment of the various components use the "Restore
Defaults" button added by the super class. Remove some intermediate
unnecessary composites. Use GridLayoutFactory and GridDataFactory
where appropriate, usually with fillDefaults() if no margin is
desired and swtDefaults() is there should be a margin. Give the
wrapping label a width hint.

Bug: 552971
Change-Id: Ic16cb08d8521ee68657ff377c1845c36a10f151e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>