Update patch set 1
Patch Set 1:

I have tried out your suggestion of adapting selectAndReveal directly. Note that if the work tree filter is active, for a selected file getNodeForPath will not be null. The node is still part of the tree's content, it just is not shown.
You would always have to add the repository node. So if the filter is *not* active, both the repository and the work tree node will be revealed (change of current behaviour). This is why I added the check for the filter.

Does this count as unwanted side-effect?

Patch-set: 1
Label: Verified=0