Update patch set 2
Patch Set 2: Code-Review-1

Looks nice to me. I forgot one thing in my last review, sorry for that. Would you mind adding the icon UIIcons.COMMIT to the new menu item? There are already 2 other images as members in SwitchToMenu, and you can do it similar.

That way it would be more consistent with the "Replace with > Commit" and "Compare with > Commit" menu items on files.

Patch-set: 2
Label: Code-Review=-1