Update patch set 2
Patch Set 2:

> I'd rather get rid of RepositoryTreeNode<?> completely. Is there a
 > real need for a separate "view model"? Why do we have a
 > RepositoryNode at all? Couldn't we use the Repository object
 > directly?

In fact, that's maybe not such a good idea. Some parts of Eclipse cache last selections (for instance the CommonNavigator, or the WorkbenchSourceProvider), so if we used Repository directly, the framework would hold onto Repository instances longer than expected. With RepositoryTreeNode, it'll only hold on to that node, but we can clear the reference to the Repository. So if one wanted to remove the RepositoryTreeNode hierarchy completely it would have to be done very carefully and might give surprising side effects.

Patch-set: 2