Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

Rebased, given the "Others..." item also an icon (looked a bit out of place without, since it was the only item that didn't have one), fixed copyright headers, and rephrased the commit message.

On the commit message: it should describe the change as a whole, not list changes made from patch set to patch set. For that, use a Gerrit comment, like I did just here.

The change is basically fine, except for that "if (itemCount > 0)". Why was that added?

Patch-set: 5