Update patch set 2
Patch Set 2: Code-Review+2

Looks good to me. I was wondering if we should create a reusable component for the "repo icon + repository name and decorator" part that we already have in some of the views. It would probably fit quite nicely at the top of the content area of many similar dialogs.

Patch-set: 2
Label: Code-Review=+2