Update patch set 7
Patch Set 7:

I had a closer look at 'CommitNonWSChanges'. The problem seems to be, that in 'clickOnCommit' it expects to find the action 'Commit...' in the context menu of the Repository view. The first item in the menu is 'Switch to...', which now also has a 'Commit...' item. That gets picked up as item to be clicked which causes the wrong dialog to show up, so the test correctly fails.
I can have a look at this tonight, I guess.
Unfortunately my 'home development machine' has the computing power of an 80s pocket calculator, so I can't run all tests at the moment. Otherwise I'd have seen that...

Patch-set: 7
Reviewer: Gerrit User 3358 <3358@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0