Update patch set 2
Patch Set 2:

Did you use the runInWorkspace encapsulation to a) make the test more stable or b) because you believe it reflects the real workflow of the unit under test? In case of the A: Assuming that your listener fixes the actual root cause of the test, a group of separate workspace operations is more challenging for the unit test. I.e. there might be many plugins in eclipse that create resources without using a workspace operation, and egit must be able to deal with that correctly.

Patch-set: 2