Update patch set 1
Patch Set 1:

> I'm just wondering about migration for nightly users.

I did think about it, but decided it was not worth the trouble. This has not been officially released yet and so I thought a migration for the few nightly
users would not be necessary. If it had been released it'd be a different story. But if you think it's necessary to auto-migrate anyway, I could add that, too.

Patch-set: 1
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0