Update patch set 1
Patch Set 1:

(2 comments)

Generally nice, but I'm really missing the N/M feedback.

Slightly confusing: the search is over the full commit message, but there's no way to see that in the dialog. So sometimes one is left wondering why a certain commit is highlighted. (Already in the history view I sometimes wonder, but there I can see the commit message. Though it would help if matches in the text were highlighted, but that's totally unrelated.)

Patch-set: 1