Update patch set 1
Patch Set 1: Code-Review+2

Code to mark the starts is basically the same as in GitHistoryWalk. Ran into this with one of my test repos, always got an empty dialog with just an error message "Object SHA-1 is not a commit". Before the ref filters were introduced, this dialog used either HEAD or HEAD, refs/heads/**, and refs/remotes/**, but not refs/tags/**. Annotated tags point to tag objects, not to commits.

And as described in the bug report, I also got simply an empty history because the filter included only a side branch that didn't have the file on which I had chosen "Compare with Commit...". Which is utter nonsense, too.

Patch-set: 1
Label: Code-Review=+2