Update patch set 2
Patch Set 2:

> Slightly confusing: the search is over the full commit message, but there's no way to see that in the dialog. So sometimes one is left wondering why a certain commit is highlighted. (Already in the history view I sometimes wonder, but there I can see the commit message. Though it would help if matches in the text were highlighted, but that's totally unrelated.)

Concerning the search extent: I agree, took me a while to understand what was in the search scope. That could be changed in FindToolbar/FindToolbarJob. I could add a new option to the already existing drop down menu that only includes the short message in the search. Maybe the existing option 'All' could be renamed to 'All (including full message)' and a new option could be 'All (including short message)'. If that's a viable solution, what would be the process to implement it? Add it to this change or create a new one(e.g. create a new bug, fix it, upload a patch set)?

Patch-set: 2
Label: Verified=0