Update patch set 1
Patch Set 1:

I didn't try it, but I'm a bit worried that doing this in the label provider on the UI thread may have noticeable performance impact.

I suggest we do this on top of https://git.eclipse.org/r/#/c/153953/ , which fixes the whole label provider/decorator setup such that we finally can do image decorations in the background decorator. Then this potentially not-so-cheap operation would also occur in the background decorator thread. (Like it does in the package/project explorer.)

Patch-set: 1
Label: Code-Review=0