[repo view] Single Ref cache for content provider and decorator
The repositories view would sometimes not update decorations correctly
on refs. This was caused by an unfortunate interaction between
decoration caching in the framework, the decorator showing information
derived from the content of a Ref, and timing issues between background
refreshes in the view and decorators firing LabelProviderChangedEvents.

Make sure that decorator and content provider both use the same Ref
cache and thus agree on the information to display.

For tags a similar problem is theoretically also possible. Instead of
also caching all tags change the equals comparison for TagNode to also
consider the content of the tag (its ObjectId). That also avoids that
the framework re-uses a cached outdated decoration, but results in a
perceptible delay between the old decoration disappearing and the new
one appearing. Since tags are moved far less frequently than other refs
this is considered acceptable.

Bug 546455
Change-Id: I7004265557ebae996b9ef10fa0d68810f1c17baf
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>