Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

Is it right that the decorator cache is reset on config change event listener, and that the view is also refreshed due to a config change event listener, but we have no guarantee this happens in the right order? The reset flag might stay "unprocessed" for some time then.

Patch-set: 3
Label: Code-Review=+1