Update patch set 3
Patch Set 3:

(1 comment)

> (1 comment)
 > 
 > Is it right that the decorator cache is reset on config change
 > event listener, and that the view is also refreshed due to a config
 > change event listener, but we have no guarantee this happens in the
 > right order? The reset flag might stay "unprocessed" for some time
 > then.

No it won't. It's correct that we don't know about the order; the important thing is that both occur. The lightweight decorators that set the flag also fire a LabelProviderChangedEvent afterwards, and that will, in case setting the flag should have occurred after the refresh() from the view, cause a re-labeling and that will pick up the flag and do the reset.

In practice I would expect setting the flag occur first, then the structural refresh from the view and the label refresh from the decorator in either order (probably the structural refresh comes last due to the longer job delay).

Patch-set: 3
Label: Verified=0